+++ ports/zephyr/ncs/src/memfault_fota.c
@@ -6,10 +6,9 @@
 #include "memfault/nrfconnect_port/http.h"
 #include "memfault/nrfconnect_port/fota.h"
 
-#include "memfault/core/debug_log.h"
-#include "memfault/core/math.h"
-#include "memfault/core/platform/core.h"
-#include "memfault/panics/assert.h"
+#include "memfault/components.h"
+
+#include "memfault/ports/ncs/version.h"
 #include "memfault/ports/zephyr/root_cert_storage.h"
 
 #include "net/download_client.h"
@@ -61,15 +60,48 @@ int memfault_fota_start(void) {
 
   // Note: The nordic FOTA API only supports passing one root CA today. So we cycle through the
   // list of required Root CAs in use by Memfault to find the appropriate one
-  const int certs[] = { kMemfaultRootCert_CyberTrustRoot, kMemfaultRootCert_DigicertRootCa,
-                        kMemfaultRootCert_DigicertRootG2, kMemfaultRootCert_AmazonRootCa1 };
+  const int certs[] = { kMemfaultRootCert_AmazonRootCa1, kMemfaultRootCert_DigicertRootCa,
+                        kMemfaultRootCert_CyberTrustRoot, kMemfaultRootCert_DigicertRootG2 };
   for (size_t i = 0; i < MEMFAULT_ARRAY_SIZE(certs); i++) {
-    rv = fota_download_start(download_url, download_url, certs[i], NULL, 0);
-    if ((rv != 0) && (errno == EOPNOTSUPP)) {
-      // We hit an error that was not due to a TLS handshake failure
-      continue;
+
+#if MEMFAULT_NCS_VERSION_GT(1, 7)
+    // In NCS 1.8 signature was changed "to accept an integer parameter specifying the PDN ID,
+    // which replaces the parameter used to specify the APN"
+    //
+    // https://github.com/nrfconnect/sdk-nrf/blob/v1.8.0/include/net/fota_download.h#L88-L106
+    rv = fota_download_start(download_url, download_url, certs[i], 0 /* pdn_id */, 0);
+#else // NCS <= 1.7
+    // https://github.com/nrfconnect/sdk-nrf/blob/v1.4.1/include/net/fota_download.h#L88-L106
+    rv = fota_download_start(download_url, download_url, certs[i], NULL /* apn */, 0);
+#endif
+    if (rv == 0) {
+      // success -- we are ready to start the FOTA download!
+      break;
     }
-    break;
+
+    if (rv == -EALREADY) {
+      MEMFAULT_LOG_INFO("fota_download_start already in progress");
+      break;
+    }
+
+    // Note: Between releases of Zephyr & NCS the error code returned for TLS handshake failures
+    // has changed.
+    //
+    // NCS < 1.6 returned EOPNOTSUPP
+    // NCS >= 1.6 returns ECONNREFUSED.
+    //
+    // The mapping of errno values has also changed across releases due to:
+    // https://github.com/zephyrproject-rtos/zephyr/commit/165def7ea6709f7f0617591f464fb95f711d2ac0
+    //
+    // Therefore we print a friendly message when an _expected_ failure takes place
+    // but retry regardless of error condition in case value returned changes again in the future.
+    if (((errno == EOPNOTSUPP) || (errno == ECONNREFUSED)) &&
+        (i != (MEMFAULT_ARRAY_SIZE(certs) - 1))) {
+      MEMFAULT_LOG_INFO("fota_download_start likely unsuccessful due to root cert mismatch. "
+                        "Trying next certificate.");
+    } else {
+      MEMFAULT_LOG_ERROR("fota_download_start unexpected failure, errno=%d", (int)errno);
+    }
   }
   if (rv != 0) {
     MEMFAULT_LOG_ERROR("FOTA start failed, rv=%d", rv);